US_Postal_Codes >> (Not for Analysis!!)
This is the master list of zip codes provided by the US Postal Service and includes 40933 rows. It contains: Zip Code, Place Name (City), State, State Abbreviation, County, Latitude, Longitude. 
Note: This is the master list of zip codes and includes extra zip codes that are not used in the analysis. These rows were filtered using the zip codes from a created metric data set, using the zip_code_manipulation.py (see in the code repository). It is only present for anyone wishing to recreate the datasets.