Made population numeric, fixed vacancy, added rankings
This is the output of the second revision of "Organize DCI data.Rmd."  It has the Population column as a numeric, not a character, field.  The vacancy percent is calculated from Mike's permanent vacancy numbers.  It has ranking columns for each of the DCI components, the average of those seven ranks, and the ranking of the averages.